ci: remove win-e2e branch test suite
