+++ .circleci/config.base.yml
@@ -122,7 +122,7 @@ jobs:
           command: |
             source .circleci/local_publish_helpers.sh
             cd packages/amplify-util-mock/
-            retry yarn e2e
+            yarn e2e
           no_output_timeout: 90m
           environment:
             JEST_JUNIT_OUTPUT: 'reports/junit/js-test-results.xml'
@@ -204,7 +204,7 @@ jobs:
           command: |
             source .circleci/local_publish_helpers.sh
             cd packages/graphql-transformers-e2e-tests/
-            retry yarn e2e --maxWorkers=3
+            yarn e2e --maxWorkers=3
           environment:
             AMPLIFY_CLI_DISABLE_LOGGING: 'true'
           no_output_timeout: 90m
@@ -315,7 +315,7 @@ jobs:
             source .circleci/local_publish_helpers.sh
             changeNpmGlobalPath
             cd packages/amplify-migration-tests
-            retry yarn run migration_v4.0.0 --maxWorkers=3 $TEST_SUITE
+            yarn run migration_v4.0.0 --maxWorkers=3 $TEST_SUITE
           no_output_timeout: 90m
       - run: *scan_e2e_test_artifacts
       - store_test_results:
@@ -422,7 +422,7 @@ jobs:
             source .circleci/local_publish_helpers.sh
             changeNpmGlobalPath
             cd packages/amplify-migration-tests
-            retry yarn run migration --maxWorkers=3 $TEST_SUITE
+            yarn run migration --maxWorkers=3 $TEST_SUITE
           no_output_timeout: 90m
       - run: *scan_e2e_test_artifacts
       - store_test_results:
