+++ packages/amplify-e2e-tests/src/__tests__/auth_1.test.ts
@@ -110,33 +110,4 @@ describe('amplify add auth...', () => {
     const userPool = await getUserPool(id, meta.providers.awscloudformation.Region);
     expect(userPool.UserPool).toBeDefined();
   });
-
-  it('...should init a project and add auth with defaults and then remove auth and add another auth and push', async () => {
-    await initJSProjectWithProfile(projRoot, defaultsSettings);
-    await addAuthWithDefault(projRoot, {});
-    await amplifyPushAuth(projRoot);
-    await removeAuthWithDefault(projRoot);
-    await addAuthWithDefault(projRoot, {});
-    await amplifyPushAuth(projRoot);
-  });
-
-  it('...should init a Flutter project and add auth with defaults', async () => {
-    await initFlutterProjectWithProfile(projRoot, defaultsSettings);
-    await addAuthWithDefault(projRoot, {});
-    await amplifyPushAuth(projRoot);
-    const meta = getProjectMeta(projRoot);
-    const id = Object.keys(meta.auth).map(key => meta.auth[key])[0].output.UserPoolId;
-    const userPool = await getUserPool(id, meta.providers.awscloudformation.Region);
-    expect(userPool.UserPool).toBeDefined();
-    expect(fs.existsSync(path.join(projRoot, 'lib', 'amplifyconfiguration.dart'))).toBe(true);
-  });
-
-  it('...should init a project and add auth with defaults and then remove auth and add another auth and push', async () => {
-    await initFlutterProjectWithProfile(projRoot, defaultsSettings);
-    await addAuthWithDefault(projRoot, {});
-    await amplifyPushAuth(projRoot);
-    await removeAuthWithDefault(projRoot);
-    await addAuthWithDefault(projRoot, {});
-    await amplifyPushAuth(projRoot);
-  });
 });
