+++ packages/amplify-e2e-tests/src/__tests__/auth_5.test.ts
@@ -115,54 +115,4 @@ describe('headless auth', () => {
     const { auth: authBackendConfigAfter } = getCloudBackendConfig(projRoot);
     expect(_.isEmpty(authBackendConfigAfter)).toBe(true);
   });
-
-  describe(' import', () => {
-    let ogProjectSettings: {name: string};
-    let ogProjectRoot: string;
-
-    beforeEach(async () => {
-      const ogProjectPrefix = 'ogauimphea';
-      ogProjectSettings = {
-        name: ogProjectPrefix,
-      };
-      ogProjectRoot = await createNewProjectDir(ogProjectSettings.name);
-      await initJSProjectWithProfile(ogProjectRoot, ogProjectSettings);
-    });
-
-    afterEach(async () => {
-      await deleteProject(ogProjectRoot);
-      deleteProjectDir(ogProjectRoot);
-    });
-
-    test.each([
-      ['userpool only', false],
-      ['userpool with identitypool', true],
-    ])(' cognito userpool %s', async (_: string, withIdentityPool: boolean) => {
-      const ogProjectDetails = await setupOgProjectWithAuth(ogProjectRoot, ogProjectSettings, withIdentityPool);
-
-      const importAuthRequest: ImportAuthRequest = {
-        version: 1,
-        userPoolId: ogProjectDetails.meta.UserPoolId,
-        nativeClientId: ogProjectDetails.meta.AppClientID,
-        webClientId: ogProjectDetails.meta.AppClientIDWeb,
-      };
-      if (withIdentityPool) {
-        importAuthRequest.identityPoolId = ogProjectDetails.meta.IdentityPoolId;
-      }
-
-      await initJSProjectWithProfile(projRoot, defaultsSettings);
-      await headlessAuthImport(projRoot, importAuthRequest);
-      await amplifyPushAuth(projRoot);
-
-      let projectDetails = getAuthProjectDetails(projRoot);
-      expectAuthProjectDetailsMatch(projectDetails, ogProjectDetails);
-      expectLocalAndCloudMetaFilesMatching(projRoot);
-
-      await removeImportedAuthWithDefault(projRoot);
-      await amplifyPushAuth(projRoot);
-
-      expectNoAuthInMeta(projRoot);
-      expectLocalTeamInfoHasNoCategories(projRoot);
-    });
-  });
 });
