+++ packages/amplify-e2e-tests/src/__tests__/layer-1.test.ts
@@ -223,198 +223,3 @@ describe('amplify add lambda layer', () => {
     await validateLayerMetadata(projRoot, settings, getProjectMeta(projRoot), envName, integtestArns);
   });
 });
-
-describe('test amplify remove function', () => {
-  let projRoot: string;
-  let projName: string;
-  const envName = 'integtest';
-
-  beforeEach(async () => {
-    projRoot = await createNewProjectDir('layers');
-    await initJSProjectWithProfile(projRoot, { envName });
-    ({ projectName: projName } = getProjectConfig(projRoot));
-  });
-
-  afterEach(async () => {
-    await deleteProject(projRoot);
-    deleteProjectDir(projRoot);
-  });
-
-  it('init a project, add layer, push 4 layer versions and delete first 3 of them, then push and verify', async () => {
-    const [shortId] = uuid().split('-');
-    const layerName = `simplelayer${shortId}`;
-    const runtime: LayerRuntime = 'nodejs';
-
-    const settings = {
-      runtimes: [runtime],
-      layerName,
-      projName,
-    };
-    const arns: string[] = [];
-    await addLayer(projRoot, settings);
-    expect(validateLayerDir(projRoot, { projName, layerName: settings.layerName }, settings.runtimes)).toBe(true);
-    await amplifyPushLayer(projRoot, {
-      acceptSuggestedLayerVersionConfigurations: true,
-    });
-    arns.push(getCurrentLayerArnFromMeta(projRoot, settings));
-    for (const i in [1, 2, 3]) {
-      updateOptData(projRoot, settings, i);
-      await amplifyPushLayer(projRoot, {
-        acceptSuggestedLayerVersionConfigurations: true,
-      });
-      arns.push(getCurrentLayerArnFromMeta(projRoot, settings));
-    }
-    const removeVersion = [1, 2, 3];
-    await removeLayerVersion(projRoot, {}, removeVersion, [1, 2, 3, 4]);
-    updateOptData(projRoot, settings, 'end');
-    await amplifyPushLayer(projRoot, {
-      acceptSuggestedLayerVersionConfigurations: true,
-    });
-    arns.push(getCurrentLayerArnFromMeta(projRoot, settings));
-    arns.splice(0, 3);
-    validateLayerMetadata(projRoot, settings, getProjectMeta(projRoot), envName, arns);
-  });
-
-  it('init a project, add layer, push 2 layer versions, add 2 dependent functions, check that removal is blocked', async () => {
-    const [shortId] = uuid().split('-');
-    const layerName = `simplelayer${shortId}`;
-    const runtime: LayerRuntime = 'nodejs';
-
-    const settings = {
-      runtimes: [runtime],
-      layerName,
-      projName,
-    };
-    const arns: string[] = [];
-    await addLayer(projRoot, settings);
-    expect(validateLayerDir(projRoot, { projName, layerName: settings.layerName }, settings.runtimes)).toBe(true);
-    await amplifyPushLayer(projRoot, { acceptSuggestedLayerVersionConfigurations: true });
-    arns.push(getCurrentLayerArnFromMeta(projRoot, settings));
-    updateOptData(projRoot, settings, 'update');
-    await amplifyPushLayer(projRoot, { acceptSuggestedLayerVersionConfigurations: true });
-    arns.push(getCurrentLayerArnFromMeta(projRoot, settings));
-
-    const fnName1 = `integtestFn1${shortId}`;
-    const fnName2 = `integtestFn2${shortId}`;
-
-    await addFunction(
-      projRoot,
-      {
-        functionTemplate: 'Hello World',
-        name: fnName1,
-        layerOptions: {
-          layerWalkthrough: (chain: ExecutionContext): void => {
-            chain
-              .wait('Provide existing layers')
-              .sendKeyDown()
-              .send(' ')
-              .sendCarriageReturn()
-              .wait(`Select a version for ${projName + layerName}`)
-              .sendKeyDown(2) // Move from Always choose latest version to version 1
-              .sendCarriageReturn();
-          },
-        },
-      },
-      runtime,
-    );
-    await addFunction(
-      projRoot,
-      {
-        functionTemplate: 'Hello World',
-        name: fnName2,
-        layerOptions: {
-          layerWalkthrough: (chain: ExecutionContext): void => {
-            chain
-              .wait('Provide existing layers')
-              .sendKeyDown()
-              .send(' ')
-              .sendCarriageReturn()
-              .wait(`Select a version for ${projName + layerName}`)
-              .sendKeyDown() // Move from Always choose latest version to version 2
-              .sendCarriageReturn();
-          },
-        },
-      },
-      runtime,
-    );
-
-    await removeLayerVersion(projRoot, { removeNoLayerVersions: true }, [1, 2], [1, 2]);
-    validateLayerMetadata(projRoot, settings, getProjectMeta(projRoot), envName, arns);
-
-    await removeFunction(projRoot, fnName1);
-    await removeFunction(projRoot, fnName2);
-
-    await removeLayerVersion(projRoot, {}, [1], [1, 2]);
-    await amplifyPushAuth(projRoot);
-    validateLayerMetadata(projRoot, settings, getProjectMeta(projRoot), envName, arns.splice(1));
-  });
-});
-
-describe('amplify add lambda layer - with amplify console app', () => {
-  let projRoot: string;
-  const envName = 'integtest';
-  let projName: string;
-
-  beforeEach(async () => {
-    projRoot = await createNewProjectDir('layers');
-    await initJSProjectWithProfile(projRoot, { envName, disableAmplifyAppCreation: false });
-    const { projectName } = getProjectConfig(projRoot);
-    projName = projectName;
-  });
-
-  afterEach(async () => {
-    await deleteProject(projRoot);
-    deleteProjectDir(projRoot);
-  });
-
-  it('tests amplify pull on project with layer', async () => {
-    const [shortId] = uuid().split('-');
-    const layerName = `testlayer${shortId}`;
-    const runtime: LayerRuntime = 'nodejs';
-    const settings = {
-      runtimes: [runtime],
-      layerName,
-      projName,
-    };
-    const expectedPerms: LayerPermission[] = [{ type: LayerPermissionName.private }];
-
-    await addLayer(projRoot, settings);
-    await amplifyPushLayer(projRoot, {
-      acceptSuggestedLayerVersionConfigurations: true,
-    });
-
-    validatePushedVersion(projRoot, { layerName, projName }, expectedPerms);
-
-    const appId = getAppId(projRoot);
-    expect(appId).toBeDefined();
-
-    let projRoot2;
-
-    try {
-      projRoot2 = await createNewProjectDir('layer-pull-test');
-      await amplifyPull(projRoot2, { override: false, emptyDir: true, appId });
-
-      validatePushedVersion(projRoot2, { layerName, projName }, expectedPerms);
-
-      // Push new resource with no change to the layer
-      await addFunction(projRoot2, { functionTemplate: 'Hello World' }, runtime);
-      await amplifyPushAuth(projRoot2);
-
-      // Push a new layer version
-      addOptData(projRoot2, settings);
-
-      await amplifyPushLayer(projRoot2, {
-        acceptSuggestedLayerVersionConfigurations: true,
-      });
-
-      validatePushedVersion(projRoot2, settings, expectedPerms);
-    } finally {
-      deleteProjectDir(projRoot2);
-    }
-
-    await amplifyPull(projRoot, {});
-    await amplifyStatus(projRoot, 'No Change');
-
-    validatePushedVersion(projRoot, settings, expectedPerms);
-  });
-});
