+++ packages/amplify-e2e-tests/src/__tests__/schema-auth-1.test.ts
@@ -23,14 +23,4 @@ describe('api directives @auth batch 1', () => {
     const testresult = await testSchema(projectDir, 'auth', 'owner2');
     expect(testresult).toBeTruthy();
   });
-
-  it('auth owner3', async () => {
-    const testresult = await testSchema(projectDir, 'auth', 'owner3');
-    expect(testresult).toBeTruthy();
-  });
-
-  it('auth owner4', async () => {
-    const testresult = await testSchema(projectDir, 'auth', 'owner4');
-    expect(testresult).toBeTruthy();
-  });
 });
