+++ packages/amplify-e2e-tests/src/__tests__/schema-auth-13.test.ts
@@ -1,11 +1,11 @@
 import { initJSProjectWithProfile, deleteProject, createNewProjectDir, deleteProjectDir } from 'amplify-e2e-core';
 import { testSchema } from '../schema-api-directives';
 
-describe('api directives @auth batch 2', () => {
+describe('api directives @auth batch 1', () => {
   let projectDir: string;
 
   beforeEach(async () => {
-    projectDir = await createNewProjectDir('auth2');
+    projectDir = await createNewProjectDir('auth1');
     await initJSProjectWithProfile(projectDir, {});
   });
 
@@ -14,12 +14,13 @@ describe('api directives @auth batch 2', () => {
     deleteProjectDir(projectDir);
   });
 
-  it('auth owner5', async () => {
-    const testresult = await testSchema(projectDir, 'auth', 'owner5');
+  it('auth owner3', async () => {
+    const testresult = await testSchema(projectDir, 'auth', 'owner3');
     expect(testresult).toBeTruthy();
   });
-  it('auth owner6', async () => {
-    const testresult = await testSchema(projectDir, 'auth', 'owner6');
+
+  it('auth owner4', async () => {
+    const testresult = await testSchema(projectDir, 'auth', 'owner4');
     expect(testresult).toBeTruthy();
   });
 });
