+++ packages/amplify-migration-tests/src/__tests__/migration_tests/transformer_migration/api.key.migration-2.test.ts
@@ -23,15 +23,20 @@ describe('amplify key force push', () => {
     deleteProjectDir(projRoot);
   });
 
-  it('init project, add key and migrate with force push', async () => {
+  it('init project, add lsi key and force push expect error', async () => {
     const projectName = 'keyforce';
-    const initialSchema = 'migrations_key/simple_key.graphql';
+    const initialSchema = 'migrations_key/initial_schema.graphql';
     // init, add api and push with installed cli
     await initJSProjectWithProfile(projRoot, { name: projectName });
     await addApiWithSchema(projRoot, initialSchema);
     await amplifyPush(projRoot);
+    // add feature flag
+    addFeatureFlag(projRoot, 'graphqltransformer', 'secondaryKeyAsGSI', true);
+    // forceUpdateSchema
+    updateApiSchema(projRoot, projectName, initialSchema, true);
     // gql-compile and force push with codebase cli
-    await apiGqlCompile(projRoot, true);
-    await amplifyPushForce(projRoot, true);
+    await expect(
+      amplifyPushUpdate(projRoot, /Attempting to remove a local secondary index on the TodoTable table in the Todo stack.*/, true),
+    ).rejects.toThrowError(/Attempting to remove a local secondary index on the TodoTable table in the Todo stack.*/);
   });
 });
